Respect language-specific cursorStyle setting
Load configuration with language scope for the language
set in the active editor.

For example, if the following is in `.vscode/settings.json`

```json
"editor.cursorStyle": "line",
"[lua]": {
    "editor.cursorStyle": "block",
}
```

then the cursor would be "line" instead of "block" in Lua files
when this extension enabled. After this commit, it has "block" style.

The language-specific scope API was added in some more recent
version of VS Code. Tried to bump up the `vscode` package version
but turns out it's deprecated so replace the package with
`@types/vscode`. Remove the no longer needed (and working)
`postinstall` script. Update lock files which changed after
`npm install`.